Add Apache license file
I updated `jira-backup` and added the capability to back up Atlassian Confluence as well.  I want to contribute my changes, but I first need a license to do so.  I see that you usually use the Apache license.

Just to make this super quick and easy for you, I have created license file `LICENSE` for you so all you need to do is merge this PR.